app, libgimp: WIP new infrastructure for testing layer modes.
One of the points we care most about is for XCF code stability over
time. In particular, loading an old XCF should render visually the same.

This unit-testing infrastructure allows testing that we do not deviate
when editing layer mode code. It works on u8 so that float computation
precision does not come into account as well.

For now, I only run it on legacy layer modes.

Note that 6 legacy modes have different hash from their 2.10
counterparts (difference, subtract and all 4 HSV/HSL layer modes). In a
next step, I will look more closely into them to determine if we are
talking of minor, not-too relevant, differences or if we broke their
implementations somehow.