app, libgimp: improve debugging ability in the layer mode unit testing.
I add a feature so that one can drop a data file of the result as
expected in the build folder. Then the unit testing framework will
compare it to the layer mode render and will list of the problematic
pixels, additionally to some statistical data to discover if we are
likely just dealing with off-by-one issues, or some more drastic render
differences.