tools: Move layer extent options down a line
Users with narrow toolbox widths may not notice
the layer extent square options as it is right next to
the "Use extents of layer content" checkbox
horizontally.
This patch adjusts the GtkGrid to move the option
directly below the checkbox, still keeping them associated
without requiring the toolbox width to be quite so large to
see it.