app, libgimp, libgimpbase: Add preference to allow saving and updating extra metadata
Closes #5856 - provide a way to export images without adding or
changing any metadata
Handles issue #3490 together with !2367. The latter makes sure that
an image comment when present is favored over similar metadata tags.
This commit makes sure that when you disable the preference to
add and update non essential metadata, it does not synchronize the
image comment with similar metadata tags (possibly overwriting other
metadata).

This adds a metadata preference (enabled by default) that on export
determines whether we add and update some non essential metadata.

Think setting GIMP as the Software used to write the metadata and
related tags, synchronizing comment with similar tags, writing save
history tags to XMP, etc.

When this setting is disabled, we only touch the metadata that we
cannot avoid (e.g. updating size, presence of thumbnail).